+++ lib/dart_lame.dart
@@ -4,5 +4,6 @@
 library dart_lame;
 
 export 'src/dart_lame_base.dart';
+export 'src/ffi/lame_loader.dart';
 
 // TODO: Export any libraries intended for clients of this package.